# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.ErrorBlock'
      message:
        type: string
      status:
        description: '"success" or "error"'
        type: string
      timestamp:
        type: string
    type: object
  github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.ErrorBlock:
    properties:
      code:
        type: string
      details: {}
    type: object
...
  internal_customer.CustomerOrderResponse:
    properties:
      aggregate_status:
        type: string
      created_at:
        type: string
      id:
        type: string
      total_price:
        type: number
    type: object
...
  internal_order.AcceptOrderRequest:
    properties:
      fulfillment_type:
        type: string
    required:
    - fulfillment_type
    type: object
  internal_order.AddTrackingRequest:
    properties:
      tracking_number:
        type: string
      tracking_url:
        type: string
    required:
    - tracking_number
    - tracking_url
    type: object
  internal_order.AddressDTO:
    properties:
      address1:
        type: string
      address2:
        type: string
      city:
        type: string
      country:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      province:
        type: string
      zip:
        type: string
    type: object
  internal_order.CancelOrderRequest:
    properties:
      fulfillment_type:
        type: string
      reason:
        type: string
    required:
    - fulfillment_type
    - reason
    type: object
  internal_order.CreateOrderRequest:
    properties:
      guest_info:
        $ref: '#/definitions/internal_order.GuestInfo'
      shipping_method:
        type: string
    required:
    - shipping_method
    type: object
  internal_order.CustomerDTO:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  internal_order.GuestInfo:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    required:
    - email
    - first_name
    - last_name
    type: object
  internal_order.LineItemsGroup:
    properties:
      pre_order:
        items:
          $ref: '#/definitions/internal_order.OrderItemDetail'
        type: array
      ship_ready:
        items:
          $ref: '#/definitions/internal_order.OrderItemDetail'
        type: array
    type: object
  internal_order.OrderItemDetail:
    properties:
      amount_charged:
        type: string
      balance_due:
        type: string
      dp_amount:
        type: string
      final_amount:
        type: string
      fulfillment_step:
        type: integer
      id:
        type: string
      item_status:
        type: string
      items_received:
        type: integer
      quantity:
        type: integer
      title:
        type: string
      type:
        type: string
      unit_price:
        type: string
      variant_id:
        type: string
    type: object
  internal_order.OrderResponse:
    properties:
      aggregate_status:
        description: deprecated
        type: string
      currency:
        type: string
      customer:
        $ref: '#/definitions/internal_order.CustomerDTO'
      financial_status:
        type: string
      fulfillment_status:
        type: string
      id:
        type: string
      line_items:
        $ref: '#/definitions/internal_order.LineItemsGroup'
      order_date:
        type: string
      order_number:
        type: string
      shipping_address:
        $ref: '#/definitions/internal_order.AddressDTO'
      shopify_checkout_url:
        type: string
      shopify_draft_invoice_url:
        type: string
      total_balance_due:
        type: string
      total_charged_now:
        type: string
      total_deposit_paid:
        type: string
      total_price:
        type: string
      total_ship_ready:
        type: string
    type: object
  internal_order.UpdateReceivedRequest:
    properties:
      items_received:
        minimum: 0
        type: integer
    required:
    - items_received
    type: object
  internal_order.UpdateStepRequest:
    properties:
      fulfillment_step:
        maximum: 4
        minimum: 1
        type: integer
    required:
    - fulfillment_step
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Momiji Home Backend API with Shopify orchestration
  termsOfService: http://swagger.io/terms/
  title: Momiji Home API
  version: "1.0"
paths:
...
  /customers/{id}/orders:
    get:
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_customer.CustomerOrderResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get customer orders
      tags:
      - Customer
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_order.OrderResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List customer orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Order Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.OrderResponse'
              type: object
      security:
      - BearerAuth: []
      - SessionAuth: []
      summary: Create an order from cart
      tags:
      - Order
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_order.OrderResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get customer order
      tags:
      - Order
  /orders/{id}/accept:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Accept Order Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.AcceptOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Accept an order
      tags:
      - Order
  /orders/{id}/cancel:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancel Order Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - Order
  /orders/{id}/items/{itemId}/received:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: Update Received Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateReceivedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update items received count
      tags:
      - Order
  /orders/{id}/items/{itemId}/step:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: Update Step Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.UpdateStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update fulfillment step
      tags:
      - Order
  /orders/{id}/items/{itemId}/tracking:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Item ID
        in: path
        name: itemId
        required: true
        type: string
      - description: Add Tracking Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_order.AddTrackingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tumbuhindigi-sys_momiji-home-backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: Add tracking number to an item
      tags:
      - Order
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionAuth:
    in: header
    name: X-Session-ID
    type: apiKey
swagger: "2.0"